구매상담게시판 글 작성시 >> board 엔터티에서 user, basket을 참조하고 그 basket이 material들 정보를 조회하도록 기능 생성중... 아직 오류 존재

+++ CMS_WebProject/src/main/java/com/example/cms_webproject/Repository/BoardRepository.java
@@ -1,7 +1,10 @@
 package com.example.cms_webproject.Repository;
 
+import com.example.cms_webproject.Model.Basket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.cms_webproject.Model.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
+
 }
+
